Added xyzt option for time dependent properties.
Note, there is a bug for mixture. It does not allow x, y, z linear interpolation.

git-svn-id: svn://136.177.114.72/svn_GW/phastpp/trunk@3523 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/phastinput/Shapefiles/Shapefile.cpp
@@ -276,7 +276,7 @@ Shapefile::Dump(std::ostream & oss)
 
 		eType = DBFGetFieldInfo(hDBF, i, szTitle, &nWidth, &nDecimals);
 		if (strlen(szTitle) > (unsigned int) nWidth)
-			panWidth[i] = strlen(szTitle);
+			panWidth[i] = (int) strlen(szTitle);
 		else
 			panWidth[i] = nWidth;
